Fixed code base, removed redundant and dead code, nested enums into associated model, and reorganized file hierarchy for ease of reading.
