+++ src/App.tsx
@@ -10,8 +10,8 @@ import { Provider } from 'react-redux';
 
 import { ToastContainer, toast } from 'react-toastify';
 import { ManagementRoutes } from './routes/Management.routes';
-import { interviewRouteArr } from './routes/Interview.routes';
 import { SurveyRoutes } from './routes/Survey.routes';
+import { InterviewRoutes } from './routes/Interview.routes';
 
 class App extends React.Component {
   public render() {
@@ -22,8 +22,9 @@ class App extends React.Component {
             <AppNav />
             <div id="main-content-container">
               <Switch>
-                  <ManagementRoutes/>
-                  <SurveyRoutes/>
+                  <Route path='/management' component={ManagementRoutes}/>
+                  <Route path='/interview' component={InterviewRoutes}/>
+                  <Route path='/surveys' component ={SurveyRoutes}/>
               </Switch>
             </div>
             <ToastContainer autoClose={2000} position={toast.POSITION.BOTTOM_RIGHT}/>
