+++ src/interview-src/component/creaInterviewComponent/createInterview.component.scss
