+++ src/interview-src/model/Associateinput.model.ts
@@ -1,13 +1,10 @@
-
-import { Interview } from "./Interview.model";
 import { InterviewFormat } from "./Interview.format.model"
 
-export interface AssociateInput {
-  associateInputId:      number,
-  formatDesc: string,
-  receivedNotifications: Date,
-  descriptionProvided: boolean,
-  interview: Interview,
-  proposedFormat: InterviewFormat
-
+export interface IAssociateInput {
+  id?: number;
+  dateNotified: Date;
+  dayNotice: boolean;
+  descriptionProvided: boolean;
+  interviewFormat: InterviewFormat | undefined;
+  proposedFormat: InterviewFormat | undefined;
 }
