--- src/interview-src/model/interviewFormat.model.ts
@@ -1,13 +1,0 @@
-
-export enum InterviewFormat {
-    none = '',
-    onSite ='On Site',
-    inPerson = 'In Person',
-    videoCall = 'Video Call',
-    phoneCall = 'Phone Call'
-}
-
-/**
- * returns all the print friendly and DB friendly representations of the formats
- */
-export const allInterviewFormats =  Object.values(InterviewFormat)
